adopt proposal to rename feature_utag to spatial_features #66
