+++ docs/source/api/tools.rst
@@ -13,7 +13,7 @@ Image Embedding
 
     tl.feature_extraction
     tl.feature_aggregation
-    tl.feature_smoothing
+    tl.spatial_features
     tl.feature_utag
 
 Tissue Geometry
