+++ src/lazyslide/tools/__init__.py
@@ -1,7 +1,7 @@
 from ._domain import spatial_domain, tile_shaper
-from ._features import feature_extraction, feature_aggregation, feature_smoothing
+from ._features import feature_extraction, feature_aggregation
 from ._signatures import RNALinker
 from ._text_annotate import text_embedding, text_image_similarity
 from ._tissue_props import tissue_props
-from ._utag import feature_utag
+from ._spatial_features  import spatial_features, feature_utag
 from ._zero_shot import zero_shot_score, slide_caption
